moved `MegaTable` into it's own file for reuse
